First attempt on settings-editor, not working!
A lot of adding to code added, but I'm so unsure when it comes to put stuff I think should be global so I made a class with a bunch of "variables'" and voids for "loading from xml" and "saving to xml".

Problem is I want the VDDPATH needs to b e global, once for settings.xml and for every script we want to call or launch. So it doesn't work as is, plus since I'm completly unsure where to put things so it's public over multiple forms/namespaces and what not.